moved utils function
